Use npm start and test
In this commit, I used `chmod +x` to make `start.sh` and
`test/basic-test.bats` executable.

I updated the hashbang/shebang in each file for accuracy and
reliability.

I'm not sure about best practices for `.bats` files. If I identify the
`bats` binary file in the `npm test` script (you can see this in
`package.json`) do I need to identify it in the hashbang in each `.bats`
file?