Issue: Added abstract member to Vehicle
Now we want to add one more method to Vehicle. It will be different for
every subclass, but we want to ensure that it is defined.

Why not add this to the interface? Well the Describe() method isn't
related to the purpose of the IAutomobile interface, so it doesn't make
sense there. We also want it to all Vehicles, not just Automobiles (aka
Bicycles).

So we make an abstract member of the Vehicle class. This means that the
entire Vehicle class must be abstract.

This causes error CS0534 because our subclasses need to implement this
new abstract method.

Errors:

Bicycle.cs(5,9): error CS0534: 'Bicycle' does not implement inherited
abstract member 'Vehicle.Describe()'
[/Users/nicholasduckwiler/Repos/prac/csharp-prac/InterfacesInheritance/InterfacesInheritance.csproj]
Truck.cs(5,11): error CS0534: 'Truck' does not implement inherited
abstract member 'Vehicle.Describe()'
[/Users/nicholasduckwiler/Repos/prac/csharp-prac/InterfacesInheritance/InterfacesInheritance.csproj]
Sedan.cs(5,11): error CS0534: 'Sedan' does not implement inherited
abstract member 'Vehicle.Describe()'
[/Users/nicholasduckwiler/Repos/prac/csharp-prac/InterfacesInheritance/InterfacesInheritance.csproj]